ui/requirements.txt
Django==4.2.4
django-cors-headers==4.2.0
boto3==1.28.30
PyYAML==6.0.1
